Change `DeleteKey` calls to use `DELETE` to avoid unexpected creation
Otherwise, an expression like
base::win::RegKey(UpdaterScopeToHKeyRoot(updater_scope),
CLIENT_STATE_KEY, Wow6432(KEY_WRITE)).DeleteKey` will create
`CLIENT_STATE_KEY` if it does not exist because of KEY_WRITE.

Fixed: 1450686
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4583363
Auto-Submit: S Ganesh <ganesh@chromium.org>
Commit-Queue: S Ganesh <ganesh@chromium.org>
Code-Coverage: Findit <findit-for-me@appspot.gserviceaccount.com>
Cr-Commit-Position: refs/heads/main@{#1152313}

+++ chrome/updater/win/installer_api.cc
@@ -100,6 +100,13 @@ void PersistLastInstallerResultValues(UpdaterScope updater_scope,
                  kRegValueLastInstallerSuccessLaunchCmdLine);
 }
 
+bool ClientStateAppKeyExists(UpdaterScope updater_scope,
+                             const std::string& app_id) {
+  return base::win::RegKey(UpdaterScopeToHKeyRoot(updater_scope),
+                           CLIENT_STATE_KEY, Wow6432(KEY_QUERY_VALUE))
+      .Valid();
+}
+
 }  // namespace
 
 InstallerOutcome::InstallerOutcome() = default;
@@ -159,6 +166,9 @@ bool SetInstallerProgressForTesting(UpdaterScope updater_scope,
 
 bool DeleteInstallerProgress(UpdaterScope updater_scope,
                              const std::string& app_id) {
+  if (!ClientStateAppKeyExists(updater_scope, app_id)) {
+    return false;
+  }
   absl::optional<base::win::RegKey> key =
       ClientStateAppKeyOpen(updater_scope, app_id, KEY_SET_VALUE);
   return key && key->DeleteValue(kRegValueInstallerProgress) == ERROR_SUCCESS;
@@ -166,6 +176,9 @@ bool DeleteInstallerProgress(UpdaterScope updater_scope,
 
 bool DeleteInstallerOutput(UpdaterScope updater_scope,
                            const std::string& app_id) {
+  if (!ClientStateAppKeyExists(updater_scope, app_id)) {
+    return false;
+  }
   absl::optional<base::win::RegKey> key = ClientStateAppKeyOpen(
       updater_scope, app_id, KEY_SET_VALUE | KEY_QUERY_VALUE);
   if (!key) {
